Setting up posthog events
- Added changes to call function easily with custom events.
- Checks if UUID stored from website
